Refaktor
