+++ src/main/kotlin/no/nav/k9/utgaende/rest/aaregv2/ArbeidsgiverOgArbeidstakerRegisterV2.kt
@@ -98,9 +98,9 @@ internal class ArbeidsgiverOgArbeidstakerRegisterV2 (
             frilansoppdrag = emptySet()
         )
 
-        val organisasjoner = json.hentOrganisasjoner(fraOgMed, tilOgMed)
-        val privateArbeidsgivere = json.hentPrivateArbeidsgivere(fraOgMed, tilOgMed)
-        val frilansoppdrag = json.hentFrilansoppdrag(fraOgMed, tilOgMed)
+        val organisasjoner = json.hentOrganisasjonerV2(fraOgMed, tilOgMed)
+        val privateArbeidsgivere = json.hentPrivateArbeidsgivereV2(fraOgMed, tilOgMed)
+        val frilansoppdrag = json.hentFrilansoppdragV2(fraOgMed, tilOgMed)
 
         return Arbeidsgivere(
             organisasjoner = organisasjoner,
@@ -108,89 +108,4 @@ internal class ArbeidsgiverOgArbeidstakerRegisterV2 (
             frilansoppdrag = frilansoppdrag
         )
     }
-}
-
-private fun JSONArray.hentFrilansoppdrag(fraOgMed: LocalDate, tilOgMed: LocalDate): Set<Frilansoppdrag> =
-    hentArbeidsgivereMedAnsettelseperiode()
-        .filter { it.erFrilansaktivitet() }
-        .map { ansettelsesforhold ->
-            val (ansattFom, ansattTom) = ansettelsesforhold.hentFomTomFraAnsettelseperiode()
-            val typeArbeidssted = ansettelsesforhold.hentArbeidsstedType()
-
-            val offentligIdent = if (typeArbeidssted == "Person") ansettelsesforhold.hentØnsketIdentFraArbeidssted("FOLKEREGISTERIDENT") else null
-            val organisasjonsnummer = if (typeArbeidssted == "Underenhet") ansettelsesforhold.hentØnsketIdentFraArbeidssted("ORGANISASJONSNUMMER") else null
-
-            Frilansoppdrag(
-                type = typeArbeidssted.somTypeArbeidssted(),
-                organisasjonsnummer = organisasjonsnummer,
-                navn = null,
-                offentligIdent = offentligIdent,
-                ansattFom = LocalDate.parse(ansattFom),
-                ansattTom = ansattTom?.let { LocalDate.parse(it) }
-            )
-        }
-        .filter { erAnsattIPerioden(it.ansattFom, it.ansattTom, fraOgMed, tilOgMed) }
-        .toSet()
-
-
-private fun JSONArray.hentPrivateArbeidsgivere(fraOgMed: LocalDate, tilOgMed: LocalDate): Set<PrivatArbeidsgiver> =
-    hentArbeidsgivereMedAnsettelseperiode()
-        .filterNot { it.erFrilansaktivitet() }
-        .filter { it.arbeidstedErPerson() }
-        .map { ansettelsesforhold ->
-            val ident = ansettelsesforhold.hentØnsketIdentFraArbeidssted("FOLKEREGISTERIDENT")
-            val (ansattFom, ansattTom) = ansettelsesforhold.hentFomTomFraAnsettelseperiode()
-
-            PrivatArbeidsgiver(
-                offentligIdent = ident,
-                ansattFom = LocalDate.parse(ansattFom),
-                ansattTom = ansattTom?.let { LocalDate.parse(it) }
-            )
-        }
-        .filter { erAnsattIPerioden(it.ansattFom, it.ansattTom, fraOgMed, tilOgMed) }
-        .distinctBy { it.offentligIdent }
-        .toSet()
-
-private fun JSONObject.hentØnsketIdentFraArbeidssted(identType: String): String {
-    val identer = this.getJSONObject("arbeidssted").getJSONArray("identer")
-    if(identer.length() == 1) return identer.getJSONObject(0).getString("ident")
-    identer.forEach {
-        it as JSONObject
-        if(it.getString("type").equals(identType)) return it.getString("ident")
-    }
-    return identer.getJSONObject(0).getString("ident")
-}
-
-private fun JSONArray.hentOrganisasjoner(fraOgMed: LocalDate, tilOgMed: LocalDate): Set<OrganisasjonArbeidsgivere> =
-    hentArbeidsgivereMedAnsettelseperiode()
-        .filterNot { it.erFrilansaktivitet() }
-        .filter { it.arbeidstedErUnderenhet() }
-        .map { ansettelsesforhold ->
-            val organisasjonsnummer = ansettelsesforhold.hentØnsketIdentFraArbeidssted("ORGANISASJONSNUMMER")
-            val (ansattFom, ansattTom) = ansettelsesforhold.hentFomTomFraAnsettelseperiode()
-
-            OrganisasjonArbeidsgivere(
-                organisasjonsnummer = organisasjonsnummer,
-                ansattFom = LocalDate.parse(ansattFom),
-                ansattTom = ansattTom?.let { LocalDate.parse(it) }
-            )
-        }
-        .filter { erAnsattIPerioden(it.ansattFom, it.ansattTom, fraOgMed, tilOgMed) }
-        .distinctBy { it.organisasjonsnummer }
-        .toSet()
-
-private fun JSONObject.erFrilansaktivitet() = getJSONObject("type").getString("kode").equals(FRILANS.type)
-private fun JSONObject.hentArbeidsstedType() = getJSONObject("arbeidssted").getString("type")
-private fun JSONObject.arbeidstedErUnderenhet() = hentArbeidsstedType().equals("Underenhet")
-private fun JSONObject.arbeidstedErPerson() = hentArbeidsstedType().equals("Person")
-
-private fun JSONArray.hentArbeidsgivereMedAnsettelseperiode(): Sequence<JSONObject> = this
-    .asSequence()
-    .map { it as JSONObject }
-    .filter { it.has("arbeidssted") }
-    .filter { it.has("ansettelsesperiode") && it.getJSONObject("ansettelsesperiode").has("startdato") }
-
-private fun JSONObject.hentFomTomFraAnsettelseperiode(): Pair<String, String?> {
-    val ansettelsesperiode = this.getJSONObject("ansettelsesperiode")
-    return Pair(ansettelsesperiode.getString("startdato"), ansettelsesperiode.getStringOrNull("sluttdato"))
 }
